Improve Reliability of RPC Clients
The BlockingConnection created by pika is disconnected by RabbitMQ if
there are no messages flowing over the connection. We can either setup
something to periodically send messages or emulate a PHP script and
create a fresh connection every request that needs one. I choose to do
a new connection each request as it simplified a lot of the code and
might remove the need for a correlation id since we can just create an
ephemeral channel that we can delete after the timeout.